requirements.txt
pypython-telegram-bot==20.6
schedule               # if using scheduled jobs
